+++ lib/cta/api/bus_tracker.rb
@@ -1,14 +1,12 @@
 # frozen_string_literal: true
 
-require "httparty"
 require "time"
 
 module CTA
   class BusTracker
-    include HTTParty
+    include CTA::API::Client
 
-    base_uri "https://www.ctabustracker.com/bustime/api/v1"
-    format :xml
+    base_uri "https://www.ctabustracker.com/bustime/api/v3/"
 
     def initialize(api_key: ENV.fetch("CTA_BUS_TRACKER_API_KEY", nil))
       @api_key = api_key
@@ -16,51 +14,66 @@ module CTA
     end
 
     def time
-      response = get("/gettime")
+      response = get("gettime")
       Time.parse(response["tm"])
     end
 
     def vehicles(vid: nil, rt: nil)
       query = build_query(vid: vid, rt: rt)
-      response = get("/getvehicles", query)
+      response = get("getvehicles", query)
       wrap_results(response["vehicle"])
     end
 
     def routes
-      response = get("/getroutes")
-      results = wrap_array(response["route"])
-      results&.to_h { |r| [r["rt"], r["rtnm"]] }
+      response = get("getroutes")
+      results = wrap_array(response["routes"])
+      results.to_h { |r| [r["rt"], r["rtnm"]] }
     end
 
     def directions(rt:)
-      response = get("/getdirections", rt: rt)
-      results = wrap_array(response["dir"])
-      results&.map { |direction| direction.split[0].downcase.to_sym }
+      response = get("getdirections", rt: rt)
+      results = wrap_array(response["directions"])
+      results.map { |direction| direction["id"].split[0].downcase.to_sym }
     end
 
     def stops(rt:, dir:)
       dir_value = dir.is_a?(Symbol) ? "#{dir} bound" : dir
-      response = get("/getstops", rt: rt, dir: dir_value)
-      wrap_results(response["stop"])
+      response = get("getstops", rt: rt, dir: dir_value)
+      wrap_results(response["stops"])
     end
 
     def patterns(pid: nil, rt: nil)
       query = {}
       query[:pid] = Array(pid).join(",") if pid
       query[:rt] = rt if rt
-      response = get("/getpatterns", query)
+      response = get("getpatterns", query)
       wrap_results(response["ptr"])
     end
 
     def predictions(stpid: nil, rt: nil, vid: nil)
       query = build_query(stpid: stpid, rt: rt, vid: vid)
-      response = get("/getpredictions", query)
+      response = get("getpredictions", query)
       wrap_results(response["prd"])
     end
 
+    def locales
+      response = get("getlocalelist")
+      wrap_results(response["locales"])
+    end
+
+    def detours(rt: nil, rtdir: nil)
+      query = {}
+      query[:rt] = Array(rt).join(",") if rt
+      query[:rtdir] = rtdir if rtdir
+      response = get("getdetours", query)
+      wrap_results(response["dtrs"])
+    end
+
     def bulletins(rt: nil, stpid: nil)
+      warn "[DEPRECATION] CTA::BusTracker#bulletins is deprecated. " \
+           "Use CTA::BusTracker#detours instead."
       query = build_query(rt: rt, stpid: stpid)
-      response = get("/getservicebulletins", query)
+      response = get("getservicebulletins", query)
       wrap_results(response["sb"])
     end
 
@@ -74,7 +87,7 @@ module CTA
         @default_key || ENV.fetch("CTA_BUS_TRACKER_API_KEY", nil)
       end
 
-      %i[time vehicles routes directions stops patterns predictions bulletins].each do |method_name|
+      %i[time vehicles routes directions stops patterns predictions locales detours bulletins].each do |method_name|
         define_method(method_name) do |**opts|
           warn "[DEPRECATION] CTA::BusTracker.#{method_name} is deprecated. " \
                "Use CTA::BusTracker.new(api_key: '...').#{method_name} instead."
@@ -86,7 +99,7 @@ module CTA
     private
 
     def get(path, extra_query = {})
-      response = self.class.get(path, query: { key: @api_key }.merge(extra_query))["bustime_response"]
+      response = http_get(path, { key: @api_key, format: "json" }.merge(extra_query))["bustime-response"]
       check_for_errors(response["error"])
       response
     end
@@ -98,22 +111,10 @@ module CTA
     end
 
     def check_for_errors(error)
-      raise CTA::API::ApiError.new(code: nil, message: error["msg"]) if error
-    end
+      return unless error
 
-    def wrap_array(object)
-      if object.nil?
-        []
-      elsif object.respond_to?(:to_ary)
-        object.to_ary || [object]
-      else
-        [object]
-      end
-    end
-
-    def wrap_results(data)
-      results = wrap_array(data)
-      results.map { |r| CTA::API::Response.new(r) } unless results.empty?
+      error = error.first if error.is_a?(Array)
+      raise CTA::API::ApiError.new(code: nil, message: error["msg"])
     end
   end
 end