+++ lib/cta/api/train_tracker.rb
@@ -1,13 +1,10 @@
 # frozen_string_literal: true
 
-require "httparty"
-
 module CTA
   class TrainTracker
-    include HTTParty
+    include CTA::API::Client
 
-    base_uri "https://lapi.transitchicago.com/api/1.0"
-    format :xml
+    base_uri "https://lapi.transitchicago.com/api/1.0/"
 
     def initialize(api_key: ENV.fetch("CTA_TRAIN_TRACKER_API_KEY", nil))
       @api_key = api_key
@@ -20,7 +17,18 @@ module CTA
       query[:mapid] = mapid if mapid
       query[:max] = max if max
 
-      response = get("/ttarrivals.aspx", query)
+      response = get("ttarrivals.aspx", query)
+      wrap_results(response["eta"])
+    end
+
+    def positions(rt:)
+      route_list = Array(rt).join(",")
+      response = get("ttpositions.aspx", rt: route_list)
+      wrap_results(response["route"])
+    end
+
+    def follow(runnumber:)
+      response = get("ttfollow.aspx", runnumber: runnumber)
       wrap_results(response["eta"])
     end
 
@@ -34,17 +42,19 @@ module CTA
         @default_key || ENV.fetch("CTA_TRAIN_TRACKER_API_KEY", nil)
       end
 
-      def arrivals(**opts)
-        warn "[DEPRECATION] CTA::TrainTracker.arrivals is deprecated. " \
-             "Use CTA::TrainTracker.new(api_key: '...').arrivals instead."
-        new(api_key: key).arrivals(**opts)
+      %i[arrivals positions follow].each do |method_name|
+        define_method(method_name) do |**opts|
+          warn "[DEPRECATION] CTA::TrainTracker.#{method_name} is deprecated. " \
+               "Use CTA::TrainTracker.new(api_key: '...').#{method_name} instead."
+          new(api_key: key).public_send(method_name, **opts)
+        end
       end
     end
 
     private
 
     def get(path, extra_query = {})
-      response = self.class.get(path, query: { key: @api_key }.merge(extra_query))["ctatt"]
+      response = http_get(path, { key: @api_key, outputType: "JSON" }.merge(extra_query))["ctatt"]
       check_for_errors(response)
       response
     end
@@ -53,21 +63,6 @@ module CTA
       return if response["errCd"] == "0"
 
       raise CTA::API::ApiError.new(code: response["errCd"], message: response["errNm"])
-    end
-
-    def wrap_array(object)
-      if object.nil?
-        []
-      elsif object.respond_to?(:to_ary)
-        object.to_ary || [object]
-      else
-        [object]
-      end
-    end
-
-    def wrap_results(data)
-      results = wrap_array(data)
-      results.map { |r| CTA::API::Response.new(r) } unless results.empty?
     end
   end
 end